requirements.txt
nose
wget
requests
distro
gitpython
